numeric months instead of long

+++ app/(drawer)/(tabs)/post/[post].tsx
@@ -321,7 +321,7 @@ export default function PostPage() {
       return "2 days ago";
     } else {
       // More than 2 days ago, show Month Day format
-      const options = { month: "long", day: "numeric" };
+      const options = { month: "numeric", day: "numeric" };
       return date.toLocaleDateString("en-US", options);
     }
   }
